Revert refactoring
This commit reverts a TypeScript migration. This change is independent
to the main change in pr/mobile/fix-back-handler. As such, it should be
included in a separate pull request.
